Move out vg locify
